Move right-click-recenter to background svg
